+++ src/algorithms/PVT/libs/serdes_monitor_pvt.h
@@ -44,15 +44,9 @@ public:
         // compatible with the version of the headers we compiled against.
         GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-        // This is a static instance of "protobuf resource cleaner" class,
-        // that would be destructed only once at the end of the program execution
-        // This prevents random crashes due to memory corruption
-        // And keeps memory leak checkers happy
-        volatile auto& protobuf_cleaner __attribute__((unused)) = protobuf_cleanup_manager::get();
-    }
-
-    ~Serdes_Monitor_Pvt()
-    {
+        // Make sure google::protobuf::ShutdownProtobufLibrary() is called only once,
+        // at the end of the program execution (see Protobuf_Cleanup_Manager)
+        Protobuf_Cleanup_Manager::get();
     }
 
     inline Serdes_Monitor_Pvt(const Serdes_Monitor_Pvt& other) noexcept : monitor_(other.monitor_)  //!< Copy constructor