+++ src/core/libs/serdes_nav_message.h
@@ -45,15 +45,9 @@ public:
         // compatible with the version of the headers we compiled against.
         GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-        // This is a static instance of "protobuf resource cleaner" class,
-        // that would be destructed only once at the end of the program execution
-        // This prevents test crashes due to memory corruption
-        // And keeps memory leak checkers happy
-        volatile auto& protobuf_cleaner __attribute__((unused)) = protobuf_cleanup_manager::get();
-    }
-
-    ~Serdes_Nav_Message()
-    {
+        // Make sure google::protobuf::ShutdownProtobufLibrary() is called only once,
+        // at the end of the program execution (see Protobuf_Cleanup_Manager)
+        Protobuf_Cleanup_Manager::get();
     }
 
     inline Serdes_Nav_Message(const Serdes_Nav_Message& other) noexcept : navmsg_(other.navmsg_)  //!< Copy constructor